requirements.txt
fastapi
pydantic
streamlit
python-multipart
faiss-cpu
opencv-python
extra_streamlit_components==0.1.60
Flask==2.3.3
numpy==1.24.3
onnxruntime==1.15.1
pandas==2.0.3
plotly==5.16.1
pymongo==4.5.0
Requests==2.31.0
scipy==1.11.2
streamlit==1.26.0
streamlit_elements==0.1.0
streamlit_extras==0.3.2
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
umap
umap-learn
shapely
descartes
seaborn
faiss-cpu
opencv-python
easydict
pymongo
python-multipart